+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- WebTools events handling [#39](https://github.com/rokwire/gateway-building-block/issues/39)
+
 [2.1.0] - 2024-02-07
 - added successteam end point
 - added successteam/pcp end point